Attempt to fix issue with display of InvalidCallbackException message.
Message was displaying the callback as 'Array' if it was an array.  This fix uses var_export to convert the array to a useful string, but seems that when the message is returned now, it displays up to the callback, & then repeats from the beginning of the callback message.  So the full message is displayed, but further downy the message than expected.